Implement the "correct" paper size
This is based on exporting a native notebook to PDF and checking the dimensions of the
result. Results may be different on reMarkable Paper Pro. Also, a lot of the apps I'm
testing seem unable to deliver content of exactly the right size, which breaks printing.
I think that it's still worth making this change but it may end up being more trouble
than it's worth.